Remove unneeded abstract base class
